+++ packages/javascript/lib/node/crypto/index.ts
@@ -6,4 +6,5 @@ export { md5Hash } from './md5Hash.js';
 export { objectId } from './objectId.js';
 export { sha1Hash } from './sha1Hash.js';
 export { sha256Hash } from './sha256Hash.js';
+export { sha512Hash } from './sha512Hash.js';
 export { numberHash } from './numberHash.js';